+++ releaf-core/lib/releaf/core_ext/array/reorder.rb
@@ -1,0 +1,5 @@
+class Array
+  def reorder(values, options)
+    Array::Reorder.call(array: self, values: values, options: options)
+  end
+end